lake/ingest/understat/firehose_transform/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
logmatic-python==0.1.7
    # via -r requirements.in
python-json-logger==2.0.2
    # via logmatic-python
